+++ src/egeo-demo/public_api.ts
@@ -9,23 +9,24 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 export { StAlertsDemoModule } from './st-alert-demo/st-alerts-demo.module';
-export { StFormDemoModule } from './st-form-demo/st-form-demo.module';
-export { StTwoListSelectionDemoModule } from './st-two-list-selection-demo/st-two-list-selection-demo.module';
-export { StTooltipDemoModule } from './st-tooltip-demo/st-tooltip-demo.module';
-export { StSwitchDemoModule } from './st-switch-demo/st-switch-demo.module';
 export { StBreadcrumbsDemoModule } from './st-breadcrumbs-demo/st-breadcrumbs-demo.module';
-export { StInputDemoModule } from './st-input-demo/st-input-demo.module';
+export { StButtonDemoModule } from './st-button-demo/st-button-demo.module';
 export { StDropdownDemoModule } from './st-dropdown-demo/st-dropdown-demo.module';
 export { StDropdownMenuDemoModule } from './st-dropdown-menu-demo/st-dropdown-menu-demo.module';
+export { StFormDemoModule } from './st-form-demo/st-form-demo.module';
 export { StHelpDemoModule } from './st-help-demo/st-help-demo.module';
 export { StInfoBoxDemoModule } from './st-info-box-demo/st-info-box-demo.module';
+export { StInputDemoModule } from './st-input-demo/st-input-demo.module';
 export { StItemListDemoModule } from './st-item-list-demo/st-item-list-demo.module';
+export { StLabelDemoModule } from './st-label-demo/index';
 export { StModalDemoModule } from './st-modal-demo/st-modal-demo.module';
-export { StPaginationDemoModule } from './st-pagination-demo/st-pagination-demo.module';
 export { StPageTitleDemoModule } from './st-page-title-demo/st-page-title-demo.module';
+export { StPaginationDemoModule } from './st-pagination-demo/st-pagination-demo.module';
 export { StSearchDemoModule } from './st-search-demo/st-search-demo.module';
 export { StSelectDemoModule } from './st-select-demo/select-demo.module';
+export { StSwitchDemoModule } from './st-switch-demo/st-switch-demo.module';
 export { StTextareaDemoModule } from './st-textarea-demo/st-textarea-demo.module';
-export { StTreeDemoModule } from './st-tree-demo/st-tree-demo.module';
 export { StTipDemoModule } from './st-tip-demo/st-tip-demo.module';
-export { StButtonDemoModule } from './st-button-demo/st-button-demo.module';
+export { StTooltipDemoModule } from './st-tooltip-demo/st-tooltip-demo.module';
+export { StTreeDemoModule } from './st-tree-demo/st-tree-demo.module';
+export { StTwoListSelectionDemoModule } from './st-two-list-selection-demo/st-two-list-selection-demo.module';