+++ src/lib/barrels.ts
@@ -26,6 +26,7 @@ import { StInfoBoxModule } from './st-info-box/st-info-box.module';
 import { StInfoCardModule } from './st-info-card/st-info-card.module';
 import { StInputModule } from './st-input/st-input.module';
 import { StItemListModule } from './st-item-list/st-item-list.module';
+import { StLabelModule } from './st-label/index';
 import { StModalModule } from './st-modal/st-modal.module';
 import { StPageTitleModule } from './st-page-title/st-page-title.module';
 import { StPaginationModule } from './st-pagination/st-pagination.module';
@@ -65,6 +66,7 @@ export const DECLARATIONS: any[] = [
    StInfoCardModule,
    StInputModule,
    StItemListModule,
+   StLabelModule,
    StModalModule,
    StPageTitleModule,
    StPaginationModule,